OA-2 Refactor Celestial object coordinates from `int` to `float`

+++ Client/Management/Server/LocalStaticGameServer.cs
@@ -47,7 +47,7 @@ namespace Engine.Management.Server
             
         }
 
-        public void AddCelestialObject(int sessionId, int objectId, int positionX, int positionY, int direction, int speed,
+        public void AddCelestialObject(int sessionId, int objectId, float positionX, float positionY, int direction, int speed,
             int classification, string name)
         {
