+++ Client/Management/Server/ScalaGameServer.cs
@@ -41,7 +41,7 @@ namespace Engine.Management.Server
             ExecuteRequest($@"/command/{sessionId}/{objectId}/{targetObjectId}/{memberId}/{targetCell}/{typeId}");
         }
 
-        public void AddCelestialObject(int sessionId, int objectId, int positionX, int positionY, int direction, int speed,
+        public void AddCelestialObject(int sessionId, int objectId, float positionX, float positionY, int direction, int speed,
             int classification, string name)
         {
